// apps/web/src/components/editor/sidebar/scriptSideBar/types/ActionTypes.ts
export enum ActionType {
  // Transform Actions (Node2D)
  MOVE_TO = 'move_to',
  MOVE_BY = 'move_by',
  TRANSLATE = 'translate',
  ROTATE_TO = 'rotate_to',
  ROTATE_BY = 'rotate_by',
  ROTATE = 'rotate',
  SCALE_TO = 'scale_to',
  SCALE_BY = 'scale_by',
  LOOK_AT = 'look_at',
  GET_ANGLE_TO = 'get_angle_to',
  TO_LOCAL = 'to_local',
  TO_GLOBAL = 'to_global',
  
  // Node Core Functions
  ADD_CHILD = 'add_child',
  REMOVE_CHILD = 'remove_child',
  QUEUE_FREE = 'queue_free',
  GET_NODE = 'get_node',
  GET_PARENT = 'get_parent',
  GET_CHILDREN = 'get_children',
  SET_PROCESS = 'set_process',
  SET_PHYSICS_PROCESS = 'set_physics_process',
  
  // Sprite2D Actions
  SET_TEXTURE = 'set_texture',
  GET_TEXTURE = 'get_texture',
  SET_REGION_ENABLED = 'set_region_enabled',
  SET_REGION_RECT = 'set_region_rect',
  SET_FLIP_H = 'set_flip_h',
  SET_FLIP_V = 'set_flip_v',
  
  // AnimatedSprite2D Actions
  PLAY_ANIMATION = 'play_animation',
  STOP_ANIMATION = 'stop_animation',
  PAUSE_ANIMATION = 'pause_animation',
  SET_ANIMATION = 'set_animation',
  GET_ANIMATION = 'get_animation',
  SET_FRAME = 'set_frame',
  GET_FRAME = 'get_frame',
  SET_SPEED_SCALE = 'set_speed_scale',
  
  // RigidBody2D Actions
  APPLY_IMPULSE = 'apply_impulse',
  APPLY_FORCE = 'apply_force',
  SET_GRAVITY_SCALE = 'set_gravity_scale',
  FREEZE_BODY = 'freeze_body',
  SET_FREEZE_ENABLED = 'set_freeze_enabled',
  SET_MASS = 'set_mass',
  SET_LINEAR_VELOCITY = 'set_linear_velocity',
  SET_ANGULAR_VELOCITY = 'set_angular_velocity',
  SET_LOCK_ROTATION = 'set_lock_rotation',
  
  // StaticBody2D Actions
  MOVE_AND_COLLIDE = 'move_and_collide',
  SET_CONSTANT_LINEAR_VELOCITY = 'set_constant_linear_velocity',
  SET_CONSTANT_ANGULAR_VELOCITY = 'set_constant_angular_velocity',
  
  // Area2D Actions
  GET_OVERLAPPING_BODIES = 'get_overlapping_bodies',
  GET_OVERLAPPING_AREAS = 'get_overlapping_areas',
  OVERLAPS_BODY = 'overlaps_body',
  OVERLAPS_AREA = 'overlaps_area',
  SET_MONITORING = 'set_monitoring',
  SET_MONITORABLE = 'set_monitorable',
  SET_GRAVITY = 'set_gravity',
  SET_GRAVITY_VECTOR = 'set_gravity_vector',
  
  // CharacterBody2D Actions
  MOVE_AND_SLIDE = 'move_and_slide',
  IS_ON_FLOOR = 'is_on_floor',
  IS_ON_CEILING = 'is_on_ceiling',
  IS_ON_WALL = 'is_on_wall',
  GET_FLOOR_NORMAL = 'get_floor_normal',
  SET_VELOCITY = 'set_velocity',
  SET_FLOOR_MAX_ANGLE = 'set_floor_max_angle',
  SET_UP_DIRECTION = 'set_up_direction',
  
  // CollisionShape2D Actions
  SET_SHAPE = 'set_shape',
  GET_SHAPE = 'get_shape',
  SET_COLLISION_DISABLED = 'set_collision_disabled',
  SET_ONE_WAY_COLLISION = 'set_one_way_collision',
  
  // CollisionPolygon2D Actions
  SET_POLYGON = 'set_polygon',
  GET_POLYGON = 'get_polygon',
  SET_BUILD_MODE = 'set_build_mode',
  
  // Property Actions
  SET_VISIBLE = 'set_visible',
  SET_MODULATE = 'set_modulate',
  SET_PROPERTY = 'set_property',
  
  // Animation Actions
  TWEEN_PROPERTY = 'tween_property',
  
  // Signal Actions
  EMIT_SIGNAL = 'emit_signal',
  CONNECT_SIGNAL = 'connect_signal',
  
  // Flow Control
  WAIT = 'wait',
  CONDITION = 'condition',
  LOOP = 'loop',
  
  // Scene Actions
  CHANGE_SCENE = 'change_scene',
  INSTANTIATE_SCENE = 'instantiate_scene',
  
  // Audio Actions
  PLAY_SOUND = 'play_sound',
  STOP_SOUND = 'stop_sound',
  
  // Input Actions
  CHECK_INPUT = 'check_input',
  
  // Debug Actions
  PRINT = 'print',
  
  // Advanced
  CUSTOM_CODE = 'custom_code'
}